evolution at two loci
Evolution of a diploid population evolving according to random sweepstakes and randomly occurring bottlenecks.  We track a two-locus genotype conferring highest selective advantage.   We are interested in the probability of fixation, and  the time to fixation conditional on fixation, of a two-locus genotype.  The loci (or sites) are taken as unlinked. 
This is the documentation  of  the   C++ code written in CWEB.