evolution at two loci
The is the CWEB source code for  simulating a diploid population evolving according to random sweepstakes and  randomly occurring bottlenecks. We are interested in the time to fixation conditional on fixation and  probability of fixation of a two-locus genotype conferring highest selective advantage.  We  are interested in how  random sweepstakes, randomly occurring bottlenecks, and  dominance mechanisms  affect the time and probability of fixation.  The problem is small and well defined, so  is suited very well for CWEB.